+++ apps/runway/src/db/db.ts
@@ -96,5 +96,6 @@ db.on('populate', async () => {
     ],
     createdAt: now,
     updatedAt: now,
+    schedule: null,
   });
 });